Move ShapewaysClient to separate assembly.
