+++ src/DynamoCLI/Program.cs
@@ -1,13 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Dynamo.Applications;
-using System.Diagnostics;
 using Dynamo.Models;
-using Dynamo.Services;
-using System.Text.RegularExpressions;
 using Dynamo.Logging;
 
 namespace DynamoCLI
