+++ src/DynamoCore/Models/DynamoModel.cs
@@ -22,7 +22,6 @@ using Dynamo.Scheduler;
 using Dynamo.Search;
 using Dynamo.Search.SearchElements;
 using Dynamo.Selection;
-using Dynamo.Services;
 using Dynamo.UpdateManager;
 using Dynamo.Utilities;
 using Dynamo.Logging;
@@ -33,6 +32,9 @@ using Greg;
 using ProtoCore;
 using ProtoCore.Exceptions;
 using ProtoCore.Runtime;
+
+using ShapewaysClient;
+
 using Compiler = ProtoAssociative.Compiler;
 // Dynamo package manager
 using Utils = Dynamo.Nodes.Utilities;
