+++ src/DynamoCore/Models/PresetModel.cs
@@ -1,12 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using Dynamo.Core;
-using Dynamo.Services;
 using System.Xml;
-using Dynamo.Nodes;
-using Dynamo.Interfaces;
 
 namespace Dynamo.Models
 {
