+++ test/Libraries/PackageManagerTests/PackageManagerTests.csproj
@@ -34,7 +34,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="Greg">
-      <HintPath>..\..\extern\greg\Greg.dll</HintPath>
+      <HintPath>..\..\..\extern\greg\Greg.dll</HintPath>
     </Reference>
     <Reference Include="Moq, Version=4.2.1507.118, Culture=neutral, PublicKeyToken=69f491c39445e920, processorArchitecture=MSIL">
       <HintPath>..\..\..\src\packages\Moq.4.2.1507.0118\lib\net40\Moq.dll</HintPath>
